requirements.txt
numpy==1.21.5
mpi4py==3.1.3
requests==2.25.1
pytest==7.1.1
pytest-cov==3.0.0
score-clustering==0.1.4
numba==0.55.1
tables==3.7.0
scipy==1.8.0
tqdm==4.63.1
h5py==3.6.0
pandas==1.4.1
recordclass==0.17.2
cufflinks==0.17.3
# googlemaps==4.4.5
matplotlib==3.5.1
plotly==4.14.3
PyYAML==6.0
responses==0.13.3
scikit-learn==1.0.2
openpyxl==3.0.9
geopy==2.2.0
xlrd==2.0.1
